+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
   available for all stream errors
 - stream: the `TokenReader` method signature has changed to make it match the
   `xmlstream.Marshaler` interface
+- stream: renamed `ErrorNS` to `NSError` for consistency
 - xmpp: the `StreamFeature` type's `Parse` function now takes an xml.Decoder
   instead of an `xml.TokenReader`
 - xmpp: IQs sent during stream negotiation are now matched against registered